modintの初期化に負の数を入れると壊れる致命的なバグを修正

+++ Utility/bigmodint.hpp
@@ -18,7 +18,7 @@ struct modint{
     uint64_t x;
 
     modint() : x(0){}
-    TT modint(T a=0) : x((a % mod + mod) % mod){}
+    TT modint(T a=0) : x((ll(a) % mod + mod) % mod){}
 
     friend mm operator+(mm a, mm b) {
         a.x += b.x;
